requirements.txt
discord.py
dotenv
requests
flask
pymongo